+++ mobile front-end/auction_mobile/lib/category_browser.dart
@@ -1,9 +1,7 @@
-import 'package:auction_mobile/components/category_viewer.dart';
-import 'package:auction_mobile/components/product_card.dart';
-import 'package:auction_mobile/product_preview.dart';
+import 'package:auction_mobile/product_card.dart';
 import 'package:flutter/material.dart';
 
-import 'components/category_card.dart';
+import 'category_card.dart';
 
 class CategoriesBrowser extends StatelessWidget {
   const CategoriesBrowser({
@@ -19,17 +17,14 @@ class CategoriesBrowser extends StatelessWidget {
     return Card(
         child: Container(
             margin: EdgeInsets.all(10),
-            height: 360,
+            height: 350,
             child: Column(children: [
+              Text(
+                'Browse by Categories',
+                style: TextStyle(fontSize: 20),
+              ),
               Container(
-                  height: 20,
-                  child: Text(
-                    'Browse by Categories',
-                    style: TextStyle(fontSize: 20),
-                  )),
-              Divider(),
-              Container(
-                  height: 50,
+                  height: 70,
                   child: TabBar(
                     isScrollable: true,
                     labelColor: Colors.black,
@@ -51,114 +46,110 @@ class CategoriesBrowser extends StatelessWidget {
                   children: [
                     Tab(
                       child: Container(
-                          height: 200,
-                          child: ListView.builder(
-                              scrollDirection: Axis.horizontal,
-                              itemCount: 7,
-                              itemBuilder: (context, index) => Hero(
-                                  tag: 'houseproduct$index',
-                                  child: Product(() {
-                                    Navigator.of(context).push(
-                                        MaterialPageRoute(
-                                            builder: (context) =>
-                                                ProductPreview(
-                                                    'houseproduct$index')));
-                                  })))),
+                        height: 200,
+                        child: ListView(
+                          scrollDirection: Axis.horizontal,
+                          children: [
+                            Product(),
+                            Product(),
+                            Product(),
+                            Product(),
+                            Product(),
+                            Product(),
+                          ],
+                        ),
+                      ),
                     ),
                     Tab(
                       child: Container(
-                          height: 200,
-                          child: ListView.builder(
-                              scrollDirection: Axis.horizontal,
-                              itemCount: 7,
-                              itemBuilder: (context, index) => Hero(
-                                  tag: 'electronicsproduct$index',
-                                  child: Product(() {
-                                    Navigator.of(context).push(MaterialPageRoute(
-                                        builder: (context) => ProductPreview(
-                                            'electronicsproduct$index')));
-                                  })))),
+                        height: 200,
+                        child: ListView(
+                          scrollDirection: Axis.horizontal,
+                          children: [
+                            Product(),
+                            Product(),
+                            Product(),
+                            Product(),
+                            Product(),
+                            Product(),
+                          ],
+                        ),
+                      ),
                     ),
                     Tab(
                       child: Container(
-                          height: 200,
-                          child: ListView.builder(
-                              scrollDirection: Axis.horizontal,
-                              itemCount: 7,
-                              itemBuilder: (context, index) => Hero(
-                                  tag: 'vehicleproduct$index',
-                                  child: Product(() {
-                                    Navigator.of(context).push(
-                                        MaterialPageRoute(
-                                            builder: (context) =>
-                                                ProductPreview(
-                                                    'vehicleproduct$index')));
-                                  })))),
+                        height: 200,
+                        child: ListView(
+                          scrollDirection: Axis.horizontal,
+                          children: [
+                            Product(),
+                            Product(),
+                            Product(),
+                            Product(),
+                            Product(),
+                            Product(),
+                          ],
+                        ),
+                      ),
                     ),
                     Tab(
                       child: Container(
-                          height: 200,
-                          child: ListView.builder(
-                              scrollDirection: Axis.horizontal,
-                              itemCount: 7,
-                              itemBuilder: (context, index) => Hero(
-                                  tag: 'governmentalproduct$index',
-                                  child: Product(() {
-                                    Navigator.of(context).push(MaterialPageRoute(
-                                        builder: (context) => ProductPreview(
-                                            'governmentalproduct$index')));
-                                  })))),
+                        height: 200,
+                        child: ListView(
+                          scrollDirection: Axis.horizontal,
+                          children: [
+                            Product(),
+                            Product(),
+                            Product(),
+                            Product(),
+                            Product(),
+                            Product(),
+                          ],
+                        ),
+                      ),
                     ),
                     Tab(
                       child: Container(
-                          height: 200,
-                          child: ListView.builder(
-                              scrollDirection: Axis.horizontal,
-                              itemCount: 7,
-                              itemBuilder: (context, index) => Hero(
-                                  tag: 'landproduct$index',
-                                  child: Product(() {
-                                    Navigator.of(context).push(
-                                        MaterialPageRoute(
-                                            builder: (context) =>
-                                                ProductPreview(
-                                                    'landproduct$index')));
-                                  })))),
+                        height: 200,
+                        child: ListView(
+                          scrollDirection: Axis.horizontal,
+                          children: [
+                            Product(),
+                            Product(),
+                            Product(),
+                            Product(),
+                            Product(),
+                            Product(),
+                          ],
+                        ),
+                      ),
                     ),
                     Tab(
                       child: Container(
-                          height: 200,
-                          child: ListView.builder(
-                              scrollDirection: Axis.horizontal,
-                              itemCount: 7,
-                              itemBuilder: (context, index) => Hero(
-                                  tag: 'rareproduct$index',
-                                  child: Product(() {
-                                    Navigator.of(context).push(
-                                        MaterialPageRoute(
-                                            builder: (context) =>
-                                                ProductPreview(
-                                                    'rareproduct$index')));
-                                  })))),
+                        height: 200,
+                        child: ListView(
+                          scrollDirection: Axis.horizontal,
+                          children: [
+                            Product(),
+                            Product(),
+                            Product(),
+                            Product(),
+                            Product(),
+                            Product(),
+                          ],
+                        ),
+                      ),
                     ),
                   ],
                 ),
               ),
-              Container(
-                  height: 50,
-                  child: Padding(
-                    padding: EdgeInsets.all(8),
-                    child: Align(
-                      alignment: Alignment.bottomRight,
-                      child: ElevatedButton(
-                        onPressed: () {
-                          Navigator.of(context).push(MaterialPageRoute(
-                              builder: (context) => CategoryViewer()));
-                        },
-                        child: Text('Show All'),
-                      ),
-                    ),
-                  ))
+              Align(
+                alignment: Alignment.bottomRight,
+                child: ElevatedButton(
+                  onPressed: () {},
+                  child: Text('Show All'),
+                ),
+              )
             ])));
   }
 }
