+++ mobile front-end/auction_mobile/lib/category_card.dart
