--- mobile front-end/auction_mobile/lib/components/notification_card.dart
@@ -1,23 +1,0 @@
-import 'package:flutter/cupertino.dart';
-import 'package:flutter/material.dart';
-
-class NotificationCard extends StatelessWidget {
-  Widget build(BuildContext context) {
-    return Card(
-      child: Container(
-        height: 100,
-        child: ListTile(
-          isThreeLine: true,
-          title: Text('Your auction is ready'),
-          subtitle: Text(
-              'Your auction is visible to customers. Check how it looks to customers'),
-          trailing: TextButton(
-            onPressed: () {},
-            child: Text('Mark as read'),
-          ),
-          onTap: () {},
-        ),
-      ),
-    );
-  }
-}
