+++ mobile front-end/auction_mobile/lib/product_browser.dart
@@ -1,9 +1,7 @@
-import 'package:auction_mobile/product_preview.dart';
 import 'package:flutter/material.dart';
 
-import 'components/category_card.dart';
-import 'components/category_viewer.dart';
-import 'components/product_card.dart';
+import 'category_card.dart';
+import 'product_card.dart';
 
 class ProductBrowser extends StatelessWidget {
   const ProductBrowser({
@@ -19,29 +17,30 @@ class ProductBrowser extends StatelessWidget {
     return Card(
         elevation: 10,
         child: Container(
-          height: 370,
+          height: 350,
           margin: EdgeInsets.all(10),
           child: Column(
             children: [
-              Container(
-                  height: 20,
-                  child: Text(
-                    'Available Products',
-                    style: TextStyle(fontSize: 20),
-                  )),
+              Text(
+                'Available Products',
+                style: TextStyle(fontSize: 20),
+              ),
               Divider(),
-              Container(
-                  height: 50,
-                  child: TabBar(
-                    labelColor: Colors.black,
-                    controller: _productsTabController,
-                    isScrollable: true,
-                    tabs: [
-                      Category(Text('Latest')),
-                      Category(Text('Popular')),
-                      Category(Text('Recommeded')),
-                    ],
-                  )),
+              TabBar(
+                labelColor: Colors.black,
+                controller: _productsTabController,
+                tabs: [
+                  Tab(
+                    child: Category(Text('Latest')),
+                  ),
+                  Tab(
+                    child: Category(Text('Popular')),
+                  ),
+                  Tab(
+                    child: Category(Text('Recommeded')),
+                  ),
+                ],
+              ),
               Container(
                 height: 200,
                 margin: EdgeInsets.all(8),
@@ -51,63 +50,59 @@ class ProductBrowser extends StatelessWidget {
                   children: [
                     Tab(
                       child: Container(
-                          height: 200,
-                          child: ListView.builder(
-                              scrollDirection: Axis.horizontal,
-                              itemCount: 7,
-                              itemBuilder: (context, index) => Hero(
-                                  tag: 'latestproduct$index',
-                                  child: Product(() {
-                                    Navigator.of(context).push(
-                                        MaterialPageRoute(
-                                            builder: (context) =>
-                                                ProductPreview(
-                                                    'latestproduct$index')));
-                                  })))),
+                        height: 200,
+                        child: ListView(
+                          scrollDirection: Axis.horizontal,
+                          children: [
+                            Product(),
+                            Product(),
+                            Product(),
+                            Product(),
+                            Product(),
+                            Product(),
+                          ],
+                        ),
+                      ),
                     ),
                     Tab(
                       child: Container(
-                          height: 200,
-                          child: ListView.builder(
-                              scrollDirection: Axis.horizontal,
-                              itemCount: 7,
-                              itemBuilder: (context, index) => Hero(
-                                  tag: 'popularproduct$index',
-                                  child: Product(() {
-                                    Navigator.of(context).push(
-                                        MaterialPageRoute(
-                                            builder: (context) =>
-                                                ProductPreview(
-                                                    'popularproduct$index')));
-                                  })))),
+                        height: 200,
+                        child: ListView(
+                          scrollDirection: Axis.horizontal,
+                          children: [
+                            Product(),
+                            Product(),
+                            Product(),
+                            Product(),
+                            Product(),
+                            Product(),
+                          ],
+                        ),
+                      ),
                     ),
                     Tab(
                       child: Container(
-                          height: 200,
-                          child: ListView.builder(
-                              scrollDirection: Axis.horizontal,
-                              itemCount: 7,
-                              itemBuilder: (context, index) => Hero(
-                                  tag: 'recommendedproduct$index',
-                                  child: Product(() {
-                                    Navigator.of(context).push(MaterialPageRoute(
-                                        builder: (context) => ProductPreview(
-                                            'recommendedproduct$index')));
-                                  })))),
+                        height: 200,
+                        child: ListView(
+                          scrollDirection: Axis.horizontal,
+                          children: [
+                            Product(),
+                            Product(),
+                            Product(),
+                            Product(),
+                            Product(),
+                            Product(),
+                          ],
+                        ),
+                      ),
                     ),
                   ],
                 ),
               ),
-              Padding(
-                  padding: EdgeInsets.all(8),
-                  child: Align(
-                      alignment: Alignment.bottomRight,
-                      child: ElevatedButton(
-                          onPressed: () {
-                            Navigator.of(context).push(MaterialPageRoute(
-                                builder: (context) => CategoryViewer()));
-                          },
-                          child: Text('Show all '))))
+              Align(
+                  alignment: Alignment.bottomRight,
+                  child: ElevatedButton(
+                      onPressed: () {}, child: Text('Show all ')))
             ],
           ),
         ));
