+++ mobile front-end/auction_mobile/lib/product_card.dart
@@ -1,0 +1,27 @@
+import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
+
+class Product extends StatelessWidget {
+  Widget build(BuildContext context) => Center(
+          child: Card(
+        elevation: 10,
+        child: SizedBox(
+          width: 200,
+          height: 200,
+          child: Column(
+            mainAxisAlignment: MainAxisAlignment.start,
+            children: [
+              SizedBox(
+                width: 200,
+                height: 130,
+                child: Image.asset('assets/images/product_logo.jpg',
+                    fit: BoxFit.cover),
+              ),
+              Divider(),
+              Text('Sample Product', style: TextStyle(fontSize: 16)),
+              Text('This is a sample ...', textAlign: TextAlign.left)
+            ],
+          ),
+        ),
+      ));
+}
